add render BackdropTeamModal, fix create and render team modal
