+++ src/js/handlers/onGoItStudentsClick.js
@@ -1,0 +1,6 @@
+import { renderTeamModal } from "../templates/render-team-modal";
+
+export default function onGoItClick() {
+    let openTeamModalBtn = document.querySelector('.footer__button');
+    openTeamModalBtn.addEventListener('click', renderTeamModal);
+}
