+++ src/js/templates/render-backdrop-teamModal.js
@@ -1,0 +1,5 @@
+export {renderBackdropTeamModal}
+    
+    function renderBackdropTeamModal() {
+    return `<div class="backdrop__team-modal is-hidden"></div>`
+}
