+++ src/partials/components/team-madal.html
@@ -1,6 +1,6 @@
-<div class="backdrop__team-modal is-hidden" data-team-modal>
+<!-- <div class="backdrop__team-modal is-hidden">
     <div class="team-modal">
-        <button class="team-modal__btn-close" data-team-modal-close type="button">
+        <button class="team-modal__btn-close" type="button">
             <svg width="11" height="11">
                 <use href="./images/sprite.svg#icon-close"></use>
             </svg>
@@ -110,6 +110,6 @@
             </div>
             </li>
         </ul>
-    </div>   -->
+    </div>
 </div>
-
+ -->
